pyproject.toml
[tool.poetry]
name = "nopay"
version = "0.1.0"
description = ""
authors = ["PashaDem <pdemeshkevich@ptsecurity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.1"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
drf-spectacular = "0.27.2"
psycopg2-binary = "2.9.9"
pyyaml = "6.0.2"
minio = "^7.2.9"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
celery = "^5.4.0"
redis = "^5.1.1"
pyocr = "^0.8.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "4.0.1"
black = "22.8.0"



[tool.poetry.group.celery.dependencies]
pyzbar = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
